split layout into several files
