+++ src/widgets/SchemaLayout.cpp
@@ -7,12 +7,9 @@
 
 #include <QClipboard>
 #include <QContextMenuEvent>
-#include <QtMath>
 #include <QMenu>
 
-#define Sqr(x) ((x)*(x))
-
-namespace {
+namespace ElementLayoutProps {
 
 QBrush getGlassBrush()
 {
@@ -71,7 +68,9 @@ const QFont& getLabelFont()
     return f;
 }
 
-} // namespace
+} // namespace ElementLayoutProps
+
+using namespace ElementLayoutProps;
 
 //------------------------------------------------------------------------------
 //                             ElementLayout
@@ -132,18 +131,6 @@ void ElementLayout::slopePainter(QPainter *painter)
 }
 
 //------------------------------------------------------------------------------
-
-#define DECLARE_ELEMENT_LAYOUT_BEGIN \
-class Layout : public ElementLayout { \
-public: \
-    Layout(Element *elem) : ElementLayout(elem) {} \
-    void paint(QPainter *painter, const QStyleOptionGraphicsItem*, QWidget*) override; \
-    void init() override;
-#define DECLARE_ELEMENT_LAYOUT_END };
-#define ELEMENT_LAYOUT_INIT void Layout::init()
-#define ELEMENT_LAYOUT_PAINT void Layout::paint(QPainter *painter, const QStyleOptionGraphicsItem*, QWidget*)
-
-//------------------------------------------------------------------------------
 namespace OpticalAxisLayout {
     DECLARE_ELEMENT_LAYOUT_BEGIN
     DECLARE_ELEMENT_LAYOUT_END
@@ -153,1158 +140,6 @@ namespace OpticalAxisLayout {
 
     ELEMENT_LAYOUT_PAINT {
         painter->drawLine(QLineF(-HW, 0, HW, 0));
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemEmptyRangeLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = _element->layoutOptions.drawNarrow ? 15 : 50;
-        HH = 5;
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        Q_UNUSED(painter)
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemMediumRangeLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = _element->layoutOptions.drawNarrow ? 15 : 30;
-        HH = 40;
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        painter->fillRect(boundingRect(), getGlassBrush());
-        painter->setPen(getGlassPen());
-        painter->drawLine(QLineF(-HW, -HH, HW, -HH));
-        painter->drawLine(QLineF(-HW, HH, HW, HH));
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemPlateLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = _element->layoutOptions.drawNarrow ? 15 : 25;
-        HH = 40;
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        painter->setPen(getGlassPen());
-        painter->setBrush(getGlassBrush());
-        painter->drawRect(boundingRect());
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemFlatMirrorLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        enum {
-            PlaceMiddle,
-            PlaceLeft,
-            PlaceRight
-        } place = PlaceMiddle;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 7; HH = 40;
-        if (!_element->owner()) return;
-        auto pos = _element->owner()->position(_element);
-        if (pos == ElementOwner::PositionAtLeft)
-            place = PlaceLeft;
-        else if (pos == ElementOwner::PositionAtRight)
-            place = PlaceRight;
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        qreal x = 0;
-        switch (place) {
-        case PlaceLeft:
-        case PlaceRight:
-            painter->setBrush(getMirrorBrush());
-            painter->setPen(getGlassPen());
-            painter->drawRect(boundingRect());
-            painter->setPen(getMirrorPen());
-            x = (place == PlaceLeft)? HW: -HW;
-            break;
-
-        case PlaceMiddle:
-            painter->setPen(getPlanePen());
-            break;
-        }
-        painter->drawLine(QLineF(x, -HH, x, HH));
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace CurvedElementLayout {
-    enum CurvedForm {
-        FormUnknown,
-        ConvexLens,         //      ()
-        ConcaveLens,        //      )(
-        ConvexConcaveLens,  //      ((
-        ConcaveConvexLens,  //      ))
-        PlanoConvexMirror,  //      |)
-        PlanoConvexLens,    //      |)
-        PlanoConcaveMirror, //      |(
-        PlanoConcaveLens,   //      |(
-        ConvexPlanoMirror,  //      (|
-        ConvexPlanoLens,    //      (|
-        ConcavePlanoMirror, //      )|
-        ConcavePlanoLens    //      )|
-    };
-
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QBrush brush;
-        CurvedForm paintMode = FormUnknown;
-        void paintCornerMark(QPainter *painter, const QString& mark) const;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        painter->setBrush(brush.style() == Qt::NoBrush ? getGlassBrush() : brush);
-        painter->setPen(getGlassPen());
-
-        const qreal ROC = 100;
-        const qreal sagitta = ROC - qSqrt(Sqr(ROC) - Sqr(HH));
-        const qreal startAngle = qRadiansToDegrees(qAsin(HH / ROC));
-        const qreal sweepAngle = 2*startAngle;
-
-        QPainterPath path;
-        QRectF rightSurface;
-        QRectF leftSurface;
-
-        switch (paintMode) {
-        case FormUnknown:
-            break;
-
-        case ConvexLens: // ()
-            rightSurface = QRectF(HW - 2*ROC, -ROC, 2*ROC, 2*ROC);
-            leftSurface = QRectF(-HW, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(HW - sagitta, HH);
-            path.arcTo(rightSurface, 360-startAngle, sweepAngle);
-            path.lineTo(-HW + sagitta, -HH);
-            path.arcTo(leftSurface, 180-startAngle, sweepAngle);
-            break;
-
-        case ConcaveLens: // )(
-            rightSurface = QRectF(HW - sagitta, -ROC, 2*ROC, 2*ROC);
-            leftSurface = QRectF(-HW + sagitta - 2*ROC, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(HW, -HH);
-            path.arcTo(rightSurface, 180-startAngle, sweepAngle);
-            path.lineTo(-HW, HH);
-            path.arcTo(leftSurface, 360-startAngle, sweepAngle);
-            break;
-
-        case ConvexConcaveLens: // ((
-            leftSurface = QRectF(-HW, -ROC, 2*ROC, 2*ROC);
-            rightSurface = QRectF(HW - sagitta, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(-HW + sagitta, -HH);
-            path.arcTo(leftSurface, 180-startAngle, sweepAngle);
-            path.lineTo(HW, HH);
-            path.arcTo(rightSurface, 180+startAngle, -sweepAngle);
-            break;
-
-        case ConcaveConvexLens: // ))
-            leftSurface = QRectF(-HW + sagitta - 2*ROC, -ROC, 2*ROC, 2*ROC);
-            rightSurface = QRectF(HW - 2*ROC, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(-HW, -HH);
-            path.arcTo(leftSurface, startAngle, -sweepAngle);
-            path.lineTo(HW - sagitta, HH);
-            path.arcTo(rightSurface, -startAngle, sweepAngle);
-            break;
-
-        case PlanoConvexMirror:
-            painter->setBrush(getMirrorBrush());
-            rightSurface = QRectF(HW - 2*ROC, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(HW - sagitta, HH);
-            path.arcTo(rightSurface, 360-startAngle, sweepAngle);
-            path.lineTo(-HW, -HH);
-            path.lineTo(-HW, HH);
-            break;
-
-        case PlanoConvexLens:
-            rightSurface = QRectF(HW - 2*ROC, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(HW - sagitta, HH);
-            path.arcTo(rightSurface, 360-startAngle, sweepAngle);
-            path.lineTo(-HW, -HH);
-            path.lineTo(-HW, HH);
-            break;
-
-        case PlanoConcaveMirror:
-            painter->setBrush(getMirrorBrush());
-            rightSurface = QRectF(HW - sagitta, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(HW, -HH);
-            path.arcTo(rightSurface, 180-startAngle, sweepAngle);
-            path.lineTo(-HW, HH);
-            path.lineTo(-HW, -HH);
-            break;
-
-        case PlanoConcaveLens:
-            rightSurface = QRectF(HW - sagitta, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(HW, -HH);
-            path.arcTo(rightSurface, 180-startAngle, sweepAngle);
-            path.lineTo(-HW, HH);
-            path.lineTo(-HW, -HH);
-            break;
-
-        case ConcavePlanoMirror:
-            painter->setBrush(getMirrorBrush());
-            leftSurface = QRectF(-HW + sagitta - 2*ROC, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(-HW, HH);
-            path.arcTo(leftSurface, 360-startAngle, sweepAngle);
-            path.lineTo(HW, -HH);
-            path.lineTo(HW, HH);
-            break;
-
-        case ConcavePlanoLens:
-            leftSurface = QRectF(-HW + sagitta - 2*ROC, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(-HW, HH);
-            path.arcTo(leftSurface, 360-startAngle, sweepAngle);
-            path.lineTo(HW, -HH);
-            path.lineTo(HW, HH);
-            break;
-
-        case ConvexPlanoMirror:
-            painter->setBrush(getMirrorBrush());
-            leftSurface = QRectF(-HW, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(-HW + sagitta, -HH);
-            path.arcTo(leftSurface, 180-startAngle, sweepAngle);
-            path.lineTo(HW, HH);
-            path.lineTo(HW, -HH);
-            break;
-
-        case ConvexPlanoLens:
-            leftSurface = QRectF(-HW, -ROC, 2*ROC, 2*ROC);
-            path.moveTo(-HW + sagitta, -HH);
-            path.arcTo(leftSurface, 180-startAngle, sweepAngle);
-            path.lineTo(HW, HH);
-            path.lineTo(HW, -HH);
-            break;
-        }
-        path.closeSubpath();
-
-        slopePainter(painter);
-        painter->drawPath(path);
-
-        switch (paintMode) {
-        case PlanoConvexMirror:
-            painter->setPen(getMirrorPen());
-            painter->drawArc(rightSurface, int(16*(360-startAngle)), int(16*sweepAngle));
-            break;
-
-        case PlanoConcaveMirror:
-            painter->setPen(getMirrorPen());
-            painter->drawArc(rightSurface, int(16*(180-startAngle)), int(16*sweepAngle));
-            break;
-
-        case ConcavePlanoMirror:
-            painter->setPen(getMirrorPen());
-            painter->drawArc(leftSurface, int(16*(360-startAngle)), int(16*sweepAngle));
-            break;
-
-        case ConvexPlanoMirror:
-            painter->setPen(getMirrorPen());
-            painter->drawArc(leftSurface, int(16*(180-startAngle)), int(16*sweepAngle));
-            break;
-
-        default: break;
-        }
-    }
-
-    void Layout::paintCornerMark(QPainter *painter, const QString& mark) const {
-        auto p = boundingRect().bottomRight();
-        p.setX(p.x()-qreal(HW)/2.0);
-        painter->setFont(getMarkTSFont());
-        painter->drawText(p, mark);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemCurveMirrorLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<CurvedElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 10; HH = 40;
-        ElemCurveMirror *mirror = dynamic_cast<ElemCurveMirror*>(_element);
-        if (!mirror || !_element->owner()) return;
-        layout.reset(new CurvedElementLayout::Layout(nullptr));
-        layout->setHalfSize(HW, HH);
-        layout->setSlope(mirror->alpha());
-        auto pos = _element->owner()->position(_element);
-        if (pos == ElementOwner::PositionAtLeft)
-            layout->paintMode = mirror->radius() > 0
-                ? CurvedElementLayout::PlanoConcaveMirror
-                : CurvedElementLayout::PlanoConvexMirror;
-        else if (pos == ElementOwner::PositionAtRight)
-            layout->paintMode = mirror->radius() > 0
-                ? CurvedElementLayout::ConcavePlanoMirror
-                : CurvedElementLayout::ConvexPlanoMirror;
-        else
-            layout->paintMode = mirror->radius() > 0
-                ? CurvedElementLayout::ConvexLens
-                : CurvedElementLayout::ConcaveLens;
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        if (layout) layout->paint(painter, nullptr, nullptr);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemThinLensLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<CurvedElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 10; HH = 40;
-        auto lens = dynamic_cast<ElemThinLens*>(_element);
-        if (!lens) return;
-        layout.reset(new CurvedElementLayout::Layout(nullptr));
-        layout->setHalfSize(HW, HH);
-        layout->setSlope(lens->alpha());
-        layout->paintMode = lens->focus() > 0
-                       ? CurvedElementLayout::ConvexLens
-                       : CurvedElementLayout::ConcaveLens;
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        if (layout) layout->paint(painter, nullptr, nullptr);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemCylinderLensTLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<CurvedElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 10; HH = 40;
-        auto lens = dynamic_cast<ElemCylinderLensT*>(_element);
-        if (!lens) return;
-        layout.reset(new CurvedElementLayout::Layout(nullptr));
-        layout->setHalfSize(HW, HH);
-        layout->setSlope(lens->alpha());
-        layout->paintMode = lens->focus() > 0
-                           ? CurvedElementLayout::ConvexLens
-                           : CurvedElementLayout::ConcaveLens;
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        if (layout) {
-            layout->paint(painter, nullptr, nullptr);
-            layout->paintCornerMark(painter, QStringLiteral("T"));
-        }
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemCylinderLensSLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<CurvedElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 10; HH = 40;
-        auto lens = dynamic_cast<ElemCylinderLensS*>(_element);
-        if (!lens) return;
-        layout.reset(new CurvedElementLayout::Layout(nullptr));
-        layout->setHalfSize(HW, HH);
-        layout->setSlope(lens->alpha());
-        layout->paintMode = lens->focus() > 0
-                ? CurvedElementLayout::ConvexLens
-                : CurvedElementLayout::ConcaveLens;
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        if (layout) {
-            layout->paint(painter, nullptr, nullptr);
-            layout->paintCornerMark(painter, QStringLiteral("S"));
-        }
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace CrystalElementLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {}
-
-    ELEMENT_LAYOUT_PAINT {
-        qreal cline = HH * qTan(qDegreesToRadians(_slopeAngle));
-
-        QPainterPath path;
-        switch (_slope)
-        {
-        case SlopeNone:
-            path.addRect(boundingRect());
-            break;
-
-        case SlopePlus:
-            path.moveTo(-HW + cline, -HH);
-            path.lineTo(HW + cline, -HH);
-            path.lineTo(HW - cline, HH);
-            path.lineTo(-HW - cline, HH);
-            path.closeSubpath();
-            break;
-
-        case SlopeMinus:
-            path.moveTo(-HW - cline, -HH);
-            path.lineTo(HW - cline, -HH);
-            path.lineTo(HW + cline, HH);
-            path.lineTo(-HW + cline, HH);
-            path.closeSubpath();
-            break;
-        }
-
-        painter->setBrush(getGlassBrush());
-        painter->setPen(getGlassPen());
-        painter->drawPath(path);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemTiltedCrystalLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<CrystalElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 15; HH = 40;
-        auto crystal = dynamic_cast<ElemTiltedCrystal*>(_element);
-        if (!crystal) return;
-        layout.reset(new CrystalElementLayout::Layout(nullptr));
-        layout->setSlope(crystal->alpha());
-        layout->setHalfSize(HW, HH);
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        if (layout) layout->paint(painter, nullptr, nullptr);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemTiltedPlateLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = _element->layoutOptions.drawNarrow ? 7 : 15;
-        HH = 40;
-        auto plate = dynamic_cast<ElemTiltedPlate*>(_element);
-        if (!plate) return;
-        setSlope(plate->alpha());
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        painter->setPen(getGlassPen());
-        painter->setBrush(getGlassBrush());
-        slopePainter(painter);
-        painter->drawRect(boundingRect());
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemBrewsterCrystalLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<CrystalElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = _element->layoutOptions.drawNarrow ? 20 : 30;
-        HH = 30;
-        layout.reset(new CrystalElementLayout::Layout(nullptr));
-        layout->setHalfSize(HW, HH);
-        layout->setSlopeAngle(40);
-        layout->setSlope(SlopePlus);
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        if (layout) layout->paint(painter, nullptr, nullptr);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemBrewsterPlateLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = _element->layoutOptions.drawNarrow ? 7 : 15;
-        HH = 40;
-        _slopeAngle = 40; _slope = SlopePlus;
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        painter->setPen(getGlassPen());
-        painter->setBrush(getGlassBrush());
-        slopePainter(painter);
-        painter->drawRect(boundingRect());
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemMatrixLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HH = 15; HW = 15;
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        painter->setPen(getGlassPen());
-        painter->fillRect(boundingRect(), getMirrorBrush());
-        painter->drawRect(QRectF(-HW, -HH, 2*HW, 2*HH));
-        painter->drawLine(QLineF(0, -HH, 0, HH));
-        painter->drawLine(QLineF(-HW, 0, HW, 0));
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemPointLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 5; HH = 5;
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        painter->setPen(getGlassPen());
-        painter->setBrush(Qt::black);
-        painter->drawEllipse(-3, -3, 6, 6);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemNormalInterfaceLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 1.5; HH = 40;
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        painter->setPen(getGlassPen());
-        painter->drawLine(QLineF(0, -HH, 0, HH));
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace InterfaceElementHeper {
-
-struct ElementPlacement {
-    enum {
-        BetweenRanges,
-        BetweenMedia,
-        BeforeMedium,
-        AfterMedium,
-    } placement;
-    enum {
-        OptionNone = 0x00,
-        OptionGrin = 0x01,
-    } options;
-};
-
-enum class ElementNeighbour {
-    Unknown,
-    Air,
-    Medium,
-    Grin,
-};
-
-ElementNeighbour getNeighbourKind(Schema* schema, Element* elem) {
-    if (elem) {
-        if (dynamic_cast<ElemGrinMedium*>(elem))
-            return ElementNeighbour::Grin;
-
-        if (dynamic_cast<ElemThermoMedium*>(elem))
-            return ElementNeighbour::Grin;
-
-        if (dynamic_cast<ElemMediumRange*>(elem))
-            return ElementNeighbour::Medium;
-
-        if (dynamic_cast<ElemEmptyRange*>(elem))
-            return ElementNeighbour::Air;
-    }
-    else if (schema->isSP())
-        return ElementNeighbour::Air;
-    return ElementNeighbour::Unknown;
-}
-
-// Check on these files how interfaces are drawn with combination with different media:
-// $$_PRO_FILE_PWD_/bin/test_files/draw_interfaces_grin.rez
-// $$_PRO_FILE_PWD_/bin/test_files/draw_interfaces_media.rez
-// $$_PRO_FILE_PWD_/bin/test_files/draw_interfaces_grin_media.rez
-ElementPlacement getPlacement(Element* elem) {
-    auto schema = dynamic_cast<Schema*>(elem->owner());
-    if (!schema) return {ElementPlacement::BetweenMedia, ElementPlacement::OptionNone};
-
-    auto left = getNeighbourKind(schema, schema->leftElement(elem));
-    auto right = getNeighbourKind(schema, schema->rightElement(elem));
-
-    if (left == ElementNeighbour::Air && right == ElementNeighbour::Air)
-        return {ElementPlacement::BetweenRanges, ElementPlacement::OptionNone};
-
-    if (left == ElementNeighbour::Medium && right == ElementNeighbour::Medium)
-        return {ElementPlacement::BetweenMedia, ElementPlacement::OptionNone};
-
-    if (left == ElementNeighbour::Grin && right == ElementNeighbour::Grin)
-        return {ElementPlacement::BetweenMedia, ElementPlacement::OptionGrin};
-
-    if (left == ElementNeighbour::Air && right == ElementNeighbour::Medium)
-        return {ElementPlacement::BeforeMedium, ElementPlacement::OptionNone};
-
-    if (left == ElementNeighbour::Air && right == ElementNeighbour::Grin)
-        return {ElementPlacement::BeforeMedium, ElementPlacement::OptionGrin};
-
-    if (left == ElementNeighbour::Medium && right == ElementNeighbour::Air)
-        return {ElementPlacement::AfterMedium, ElementPlacement::OptionNone};
-
-    if (left == ElementNeighbour::Grin && right == ElementNeighbour::Air)
-        return {ElementPlacement::AfterMedium, ElementPlacement::OptionGrin};
-
-    return {ElementPlacement::BetweenMedia, ElementPlacement::OptionNone};
-}
-
-}
-
-using namespace InterfaceElementHeper;
-
-//------------------------------------------------------------------------------
-// Base layout for ElemBrewsterInterfaceLayout and ElemTiltedInterfaceLayout interface.
-// Normal interface is painted by separate and simpler ElemNormalInterfaceLayout.
-namespace PlaneInterfaceElementLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        ElementPlacement placement;
-        QBrush getBrush() const;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        placement = getPlacement(_element);
-    }
-
-    QBrush Layout::getBrush() const {
-        if (placement.options & ElementPlacement::OptionGrin)
-            return getGrinBrush(HH);
-        return getGlassBrush();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        switch (placement.placement) {
-        case ElementPlacement::BetweenRanges:
-            painter->setPen(getPlanePen());
-            painter->drawLine(QLineF(HW, -HH, -HW, HH));
-            break;
-
-        case ElementPlacement::BetweenMedia:
-            painter->fillRect(boundingRect(), getBrush());
-            painter->setPen(getGlassPen());
-            painter->drawLine(QLineF(-HW, -HH, HW, -HH));
-            painter->drawLine(QLineF(HW, -HH, -HW, HH));
-            painter->drawLine(QLineF(-HW, HH, HW, HH));
-            break;
-
-        case ElementPlacement::AfterMedium: {
-                QPainterPath path;
-                path.moveTo(-HW, -HH);
-                path.lineTo(HW, -HH);
-                path.lineTo(-HW, HH);
-                path.closeSubpath();
-                painter->fillPath(path, getBrush());
-
-                painter->setPen(getGlassPen());
-                painter->drawLine(QLineF(-HW, -HH, HW, -HH));
-                painter->drawLine(QLineF(HW, -HH, -HW, HH));
-            }
-            break;
-
-        case ElementPlacement::BeforeMedium: {
-                QPainterPath path;
-                path.moveTo(HW, -HH);
-                path.lineTo(-HW, HH);
-                path.lineTo(HW, HH);
-                path.closeSubpath();
-                painter->fillPath(path, getBrush());
-
-                painter->setPen(getGlassPen());
-                painter->drawLine(QLineF(HW, -HH, -HW, HH));
-                painter->drawLine(QLineF(-HW, HH, HW, HH));
-            }
-            break;
-        }
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemBrewsterInterfaceLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<PlaneInterfaceElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 30; HH = 40;
-        layout.reset(new PlaneInterfaceElementLayout::Layout(_element));
-        layout->setHalfSize(HW, HH);
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        layout->paint(painter, nullptr, nullptr);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemTiltedInterfaceLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<PlaneInterfaceElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 15; HH = 40;
-        layout.reset(new PlaneInterfaceElementLayout::Layout(_element));
-        layout->setHalfSize(HW, HH);
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        layout->paint(painter, nullptr, nullptr);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemSphericalInterfaceLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    private:
-        InterfaceElementHeper::ElementPlacement _placement;
-        QRectF _surface;
-        qreal _startAngle;
-        qreal _sweepAngle;
-        QBrush getBrush() const;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    QBrush Layout::getBrush() const {
-        if (_placement.options & ElementPlacement::OptionGrin)
-            return getGrinBrush(HH);
-        return getGlassBrush();
-    }
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 15; HH = 40;
-        _placement = InterfaceElementHeper::getPlacement(_element);
-        auto intf = dynamic_cast<ElemSphericalInterface*>(_element);
-        const qreal ROC = 100;
-        const qreal sagitta = ROC - qSqrt(Sqr(ROC) - Sqr(HH));
-        const qreal startAngle = qRadiansToDegrees(qAsin(HH / ROC));
-        if (intf && intf->radius() < 0) {
-            _surface = QRectF(-sagitta, -ROC, 2*ROC, 2*ROC);
-            _startAngle = 180 - startAngle;
-            _sweepAngle = 2*startAngle;
-        }
-        else {
-            _surface = QRectF(sagitta - 2*ROC, -ROC, 2*ROC, 2*ROC);
-            _startAngle = startAngle;
-            _sweepAngle = -2*startAngle;
-        }
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        switch (_placement.placement) {
-        case ElementPlacement::BetweenRanges:
-            painter->setPen(getPlanePen());
-            painter->drawArc(_surface, int(_startAngle)*16, int(_sweepAngle)*16);
-            break;
-
-        case ElementPlacement::BetweenMedia:
-            painter->fillRect(boundingRect(), getBrush());
-            painter->setPen(getGlassPen());
-            painter->drawLine(QLineF(-HW, -HH, HW, -HH));
-            painter->drawLine(QLineF(-HW, HH, HW, HH));
-            painter->drawArc(_surface, int(_startAngle)*16, int(_sweepAngle)*16);
-            break;
-
-        case ElementPlacement::AfterMedium: {
-                QPainterPath path;
-                path.moveTo(-HW, -HH);
-                path.lineTo(0, -HH);
-                path.arcTo(_surface, _startAngle, _sweepAngle);
-                path.lineTo(-HW, HH);
-                path.closeSubpath();
-                painter->fillPath(path, getBrush());
-
-                painter->setPen(getGlassPen());
-                QPainterPath path1;
-                path1.moveTo(-HW, -HH);
-                path1.lineTo(0, -HH);
-                path1.arcTo(_surface, _startAngle, _sweepAngle);
-                path1.lineTo(-HW, HH);
-                painter->drawPath(path1);
-            }
-            break;
-
-        case ElementPlacement::BeforeMedium: {
-                QPainterPath path;
-                path.moveTo(HW, -HH);
-                path.lineTo(0, -HH);
-                path.arcTo(_surface, _startAngle, _sweepAngle);
-                path.lineTo(HW, HH);
-                path.closeSubpath();
-                painter->fillPath(path, getBrush());
-
-                painter->setPen(getGlassPen());
-                QPainterPath path1;
-                path1.moveTo(HW, -HH);
-                path1.lineTo(0, -HH);
-                path1.arcTo(_surface, _startAngle, _sweepAngle);
-                path1.lineTo(HW, HH);
-                painter->drawPath(path1);
-            }
-            break;
-        }
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemThickLensLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<CurvedElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = _element->layoutOptions.drawNarrow ? 10 : 25;
-        HH = 40;
-        auto lens = dynamic_cast<ElemThickLens*>(_element);
-        if (!lens) return;
-        layout.reset(new CurvedElementLayout::Layout(nullptr));
-        layout->setHalfSize(HW, HH);
-        auto R1 = lens->radius1();
-        auto R2 = lens->radius2();
-        if (R1 < 0 && R2 > 0) // ()
-            layout->paintMode = CurvedElementLayout::ConvexLens;
-        else if (R1 > 0 && R2 < 0) // )(
-            layout->paintMode = CurvedElementLayout::ConcaveLens;
-        else if (R1 < 0 && R2 < 0) // ((
-            layout->paintMode = CurvedElementLayout::ConvexConcaveLens;
-        else if (R1 > 0 && R2 > 0) // ))
-            layout->paintMode = CurvedElementLayout::ConcaveConvexLens;
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        if (layout) layout->paint(painter, nullptr, nullptr);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemGrinLensLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = _element->layoutOptions.drawNarrow ? 15 : 25;
-        HH = 40;
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        QPainterPath path;
-        path.addRect(boundingRect());
-        painter->setBrush(getGrinBrush(HH));
-        painter->setPen(getGlassPen());
-        painter->drawPath(path);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemGrinMediumLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = _element->layoutOptions.drawNarrow ? 15 : 30;
-        HH = 40;
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        painter->fillRect(boundingRect(), getGrinBrush(HH));
-        painter->setPen(getGlassPen());
-        painter->drawLine(QLineF(-HW, -HH, HW, -HH));
-        painter->drawLine(QLineF(-HW, HH, HW, HH));
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemGaussApertureLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 5;
-        HH = 40;
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        QPainterPath path;
-        path.addRect(boundingRect());
-        painter->setBrush(getGrinBrush(HH));
-        painter->setPen(getGlassPen());
-        painter->drawPath(path);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemGaussApertureLensLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<CurvedElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 10; HH = 40;
-        auto lens = dynamic_cast<ElemGaussApertureLens*>(_element);
-        if (!lens) return;
-        layout.reset(new CurvedElementLayout::Layout(nullptr));
-        layout->setHalfSize(HW, HH);
-        layout->brush = getGrinBrush(HH);
-        layout->paintMode = lens->focusT() > 0
-                       ? CurvedElementLayout::PlanoConvexLens // |)
-                       : CurvedElementLayout::PlanoConcaveLens; // |(
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        if (layout) layout->paint(painter, nullptr, nullptr);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace AxiconElementLayout {
-    enum AxiconForm {
-        FormUnknown,
-        ConvexLens,         //      <>
-        ConcaveLens,        //      ><
-        PlanoConvexMirror,  //      |>
-        PlanoConvexLens,    //      |>
-        PlanoConcaveMirror, //      |<
-        PlanoConcaveLens,   //      |<
-        ConvexPlanoMirror,  //      <|
-        ConcavePlanoMirror, //      >|
-    };
-
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        AxiconForm paintMode = FormUnknown;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        const qreal sagitta = HW * 0.85;
-
-        painter->setPen(getGlassPen());
-
-        switch (paintMode) {
-        case FormUnknown:
-            break;
-        case ConvexLens: // <>
-        case ConcaveLens: // )(
-        case PlanoConvexLens: // |>
-        case PlanoConcaveLens: // |<
-            painter->setBrush(getGlassBrush());
-            break;
-        case PlanoConvexMirror: // |>
-        case PlanoConcaveMirror: // |<
-        case ConvexPlanoMirror:  // <|
-        case ConcavePlanoMirror: // >|
-            painter->setBrush(getMirrorBrush());
-            break;
-        }
-
-        QPainterPath path;
-
-        switch (paintMode) {
-        case FormUnknown:
-            break;
-
-        case ConvexLens: // <>
-            path.moveTo(HW, 0);
-            path.lineTo(HW - sagitta, HH);
-            path.lineTo(-HW + sagitta, HH);
-            path.lineTo(-HW, 0);
-            path.lineTo(-HW + sagitta, -HH);
-            path.lineTo(HW - sagitta, -HH);
-            break;
-
-        case ConcaveLens: // )(
-            path.moveTo(HW - sagitta, 0);
-            path.lineTo(HW, HH);
-            path.lineTo(-HW, HH);
-            path.lineTo(-HW + sagitta, 0);
-            path.lineTo(-HW, -HH);
-            path.lineTo(HW, -HH);
-            break;
-
-        case PlanoConvexMirror: // |>
-        case PlanoConvexLens: // |>
-            path.moveTo(-HW * 0.5, HH);
-            path.lineTo(HW - sagitta, HH);
-            path.lineTo(HW, 0);
-            path.lineTo(HW - sagitta, -HH);
-            path.lineTo(-HW * 0.5, -HH);
-            break;
-
-        case PlanoConcaveMirror: // |<
-        case PlanoConcaveLens: // |<
-            path.moveTo(0, 0);
-            path.lineTo(HW, HH);
-            path.lineTo(-HW * 0.5, HH);
-            path.lineTo(-HW * 0.5, -HH);
-            path.lineTo(HW, -HH);
-            break;
-
-        case ConvexPlanoMirror:  // <|
-            path.moveTo(HW * 0.5, HH);
-            path.lineTo(-HW + sagitta, HH);
-            path.lineTo(-HW, 0);
-            path.lineTo(-HW + sagitta, -HH);
-            path.lineTo(HW * 0.5, -HH);
-            break;
-
-        case ConcavePlanoMirror: // >|
-            path.moveTo(HW * 0.5, HH);
-            path.lineTo(-HW, HH);
-            path.lineTo(0, 0);
-            path.lineTo(-HW, -HH);
-            path.lineTo(HW * 0.5, -HH);
-            break;
-        }
-        path.closeSubpath();
-
-        slopePainter(painter);
-        painter->drawPath(path);
-
-        painter->setPen(getMirrorPen());
-        painter->setBrush(Qt::NoBrush);
-        QPainterPath path1;
-        switch (paintMode) {
-        case PlanoConvexMirror: // |>
-            path1.moveTo(HW - sagitta, HH);
-            path1.lineTo(HW, 0);
-            path1.lineTo(HW - sagitta, -HH);
-            break;
-
-        case PlanoConcaveMirror: // |<
-            path1.moveTo(HW, HH);
-            path1.lineTo(HW - sagitta, 0);
-            path1.lineTo(HW, -HH);
-            break;
-
-        case ConvexPlanoMirror:  // <|
-            path1.moveTo(-HW + sagitta, HH);
-            path1.lineTo(-HW, 0);
-            path1.lineTo(-HW + sagitta, -HH);
-            break;
-
-        case ConcavePlanoMirror: // >|
-            path1.moveTo(-HW, HH);
-            path1.lineTo(-HW + sagitta, 0);
-            path1.lineTo(-HW, -HH);
-            break;
-
-        default: break;
-        }
-        painter->drawPath(path1);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemAxiconMirrorLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<AxiconElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 12; HH = 40;
-        auto mirror = dynamic_cast<ElemAxiconMirror*>(_element);
-        if (!mirror || !_element->owner()) return;
-        layout.reset(new AxiconElementLayout::Layout(nullptr));
-        layout->setHalfSize(HW, HH);
-        layout->setSlope(mirror->alpha());
-        auto pos = _element->owner()->position(_element);
-        if (pos == ElementOwner::PositionAtLeft)
-            layout->paintMode = mirror->theta() > 0
-                ? AxiconElementLayout::PlanoConcaveMirror
-                : AxiconElementLayout::PlanoConvexMirror;
-        else if (pos == ElementOwner::PositionAtRight)
-            layout->paintMode = mirror->theta() > 0
-                ? AxiconElementLayout::ConcavePlanoMirror
-                : AxiconElementLayout::ConvexPlanoMirror;
-        else
-            layout->paintMode = mirror->theta() > 0
-                ? AxiconElementLayout::ConvexLens
-                : AxiconElementLayout::ConcaveLens;
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        if (layout) layout->paint(painter, nullptr, nullptr);
-    }
-}
-
-//------------------------------------------------------------------------------
-namespace ElemAxiconLensLayout {
-    DECLARE_ELEMENT_LAYOUT_BEGIN
-        QSharedPointer<AxiconElementLayout::Layout> layout;
-    DECLARE_ELEMENT_LAYOUT_END
-
-    ELEMENT_LAYOUT_INIT {
-        HW = 12; HH = 40;
-        auto mirror = dynamic_cast<ElemAxiconLens*>(_element);
-        if (!mirror || !_element->owner()) return;
-        layout.reset(new AxiconElementLayout::Layout(nullptr));
-        layout->setHalfSize(HW, HH);
-        layout->setSlope(mirror->alpha());
-        auto pos = _element->owner()->position(_element);
-        if (pos == ElementOwner::PositionAtLeft)
-            layout->paintMode = mirror->theta() > 0
-                ? AxiconElementLayout::PlanoConcaveMirror
-                : AxiconElementLayout::PlanoConvexMirror;
-        else if (pos == ElementOwner::PositionAtRight)
-            layout->paintMode = mirror->theta() > 0
-                ? AxiconElementLayout::ConcavePlanoMirror
-                : AxiconElementLayout::ConvexPlanoMirror;
-        else
-            layout->paintMode = mirror->theta() > 0
-                ? AxiconElementLayout::PlanoConvexLens
-                : AxiconElementLayout::PlanoConcaveLens;
-        layout->init();
-    }
-
-    ELEMENT_LAYOUT_PAINT {
-        if (layout) layout->paint(painter, nullptr, nullptr);
     }
 }
 
@@ -1456,6 +291,12 @@ void SchemaLayout::copyImage()
 //------------------------------------------------------------------------------
 //                             ElementLayoutFactory
 //------------------------------------------------------------------------------
+
+#include "SchemaLayoutElems.h"
+#include "SchemaLayoutLens.h"
+#include "SchemaLayoutCrystal.h"
+#include "SchemaLayoutAxicon.h"
+#include "SchemaLayoutIfaces.h"
 
 namespace ElementLayoutFactory {
