+++ src/widgets/SchemaLayoutAxicon.h
@@ -1,0 +1,216 @@
+#ifndef SCHEMA_LAYOUT_AXICON_H
+#define SCHEMA_LAYOUT_AXICON_H
+
+#include "SchemaLayout.h"
+
+using namespace ElementLayoutProps;
+
+//------------------------------------------------------------------------------
+namespace AxiconElementLayout {
+    enum AxiconForm {
+        FormUnknown,
+        ConvexLens,         //      <>
+        ConcaveLens,        //      ><
+        PlanoConvexMirror,  //      |>
+        PlanoConvexLens,    //      |>
+        PlanoConcaveMirror, //      |<
+        PlanoConcaveLens,   //      |<
+        ConvexPlanoMirror,  //      <|
+        ConcavePlanoMirror, //      >|
+    };
+
+    DECLARE_ELEMENT_LAYOUT_BEGIN
+        AxiconForm paintMode = FormUnknown;
+    DECLARE_ELEMENT_LAYOUT_END
+
+    ELEMENT_LAYOUT_INIT {
+    }
+
+    ELEMENT_LAYOUT_PAINT {
+        const qreal sagitta = HW * 0.85;
+
+        painter->setPen(getGlassPen());
+
+        switch (paintMode) {
+        case FormUnknown:
+            break;
+        case ConvexLens: // <>
+        case ConcaveLens: // )(
+        case PlanoConvexLens: // |>
+        case PlanoConcaveLens: // |<
+            painter->setBrush(getGlassBrush());
+            break;
+        case PlanoConvexMirror: // |>
+        case PlanoConcaveMirror: // |<
+        case ConvexPlanoMirror:  // <|
+        case ConcavePlanoMirror: // >|
+            painter->setBrush(getMirrorBrush());
+            break;
+        }
+
+        QPainterPath path;
+
+        switch (paintMode) {
+        case FormUnknown:
+            break;
+
+        case ConvexLens: // <>
+            path.moveTo(HW, 0);
+            path.lineTo(HW - sagitta, HH);
+            path.lineTo(-HW + sagitta, HH);
+            path.lineTo(-HW, 0);
+            path.lineTo(-HW + sagitta, -HH);
+            path.lineTo(HW - sagitta, -HH);
+            break;
+
+        case ConcaveLens: // )(
+            path.moveTo(HW - sagitta, 0);
+            path.lineTo(HW, HH);
+            path.lineTo(-HW, HH);
+            path.lineTo(-HW + sagitta, 0);
+            path.lineTo(-HW, -HH);
+            path.lineTo(HW, -HH);
+            break;
+
+        case PlanoConvexMirror: // |>
+        case PlanoConvexLens: // |>
+            path.moveTo(-HW * 0.5, HH);
+            path.lineTo(HW - sagitta, HH);
+            path.lineTo(HW, 0);
+            path.lineTo(HW - sagitta, -HH);
+            path.lineTo(-HW * 0.5, -HH);
+            break;
+
+        case PlanoConcaveMirror: // |<
+        case PlanoConcaveLens: // |<
+            path.moveTo(0, 0);
+            path.lineTo(HW, HH);
+            path.lineTo(-HW * 0.5, HH);
+            path.lineTo(-HW * 0.5, -HH);
+            path.lineTo(HW, -HH);
+            break;
+
+        case ConvexPlanoMirror:  // <|
+            path.moveTo(HW * 0.5, HH);
+            path.lineTo(-HW + sagitta, HH);
+            path.lineTo(-HW, 0);
+            path.lineTo(-HW + sagitta, -HH);
+            path.lineTo(HW * 0.5, -HH);
+            break;
+
+        case ConcavePlanoMirror: // >|
+            path.moveTo(HW * 0.5, HH);
+            path.lineTo(-HW, HH);
+            path.lineTo(0, 0);
+            path.lineTo(-HW, -HH);
+            path.lineTo(HW * 0.5, -HH);
+            break;
+        }
+        path.closeSubpath();
+
+        slopePainter(painter);
+        painter->drawPath(path);
+
+        painter->setPen(getMirrorPen());
+        painter->setBrush(Qt::NoBrush);
+        QPainterPath path1;
+        switch (paintMode) {
+        case PlanoConvexMirror: // |>
+            path1.moveTo(HW - sagitta, HH);
+            path1.lineTo(HW, 0);
+            path1.lineTo(HW - sagitta, -HH);
+            break;
+
+        case PlanoConcaveMirror: // |<
+            path1.moveTo(HW, HH);
+            path1.lineTo(HW - sagitta, 0);
+            path1.lineTo(HW, -HH);
+            break;
+
+        case ConvexPlanoMirror:  // <|
+            path1.moveTo(-HW + sagitta, HH);
+            path1.lineTo(-HW, 0);
+            path1.lineTo(-HW + sagitta, -HH);
+            break;
+
+        case ConcavePlanoMirror: // >|
+            path1.moveTo(-HW, HH);
+            path1.lineTo(-HW + sagitta, 0);
+            path1.lineTo(-HW, -HH);
+            break;
+
+        default: break;
+        }
+        painter->drawPath(path1);
+    }
+}
+
+//------------------------------------------------------------------------------
+namespace ElemAxiconMirrorLayout {
+    DECLARE_ELEMENT_LAYOUT_BEGIN
+        QSharedPointer<AxiconElementLayout::Layout> layout;
+    DECLARE_ELEMENT_LAYOUT_END
+
+    ELEMENT_LAYOUT_INIT {
+        HW = 12; HH = 40;
+        auto mirror = dynamic_cast<ElemAxiconMirror*>(_element);
+        if (!mirror || !_element->owner()) return;
+        layout.reset(new AxiconElementLayout::Layout(nullptr));
+        layout->setHalfSize(HW, HH);
+        layout->setSlope(mirror->alpha());
+        auto pos = _element->owner()->position(_element);
+        if (pos == ElementOwner::PositionAtLeft)
+            layout->paintMode = mirror->theta() > 0
+                ? AxiconElementLayout::PlanoConcaveMirror
+                : AxiconElementLayout::PlanoConvexMirror;
+        else if (pos == ElementOwner::PositionAtRight)
+            layout->paintMode = mirror->theta() > 0
+                ? AxiconElementLayout::ConcavePlanoMirror
+                : AxiconElementLayout::ConvexPlanoMirror;
+        else
+            layout->paintMode = mirror->theta() > 0
+                ? AxiconElementLayout::ConvexLens
+                : AxiconElementLayout::ConcaveLens;
+        layout->init();
+    }
+
+    ELEMENT_LAYOUT_PAINT {
+        if (layout) layout->paint(painter, nullptr, nullptr);
+    }
+}
+
+//------------------------------------------------------------------------------
+namespace ElemAxiconLensLayout {
+    DECLARE_ELEMENT_LAYOUT_BEGIN
+        QSharedPointer<AxiconElementLayout::Layout> layout;
+    DECLARE_ELEMENT_LAYOUT_END
+
+    ELEMENT_LAYOUT_INIT {
+        HW = 12; HH = 40;
+        auto mirror = dynamic_cast<ElemAxiconLens*>(_element);
+        if (!mirror || !_element->owner()) return;
+        layout.reset(new AxiconElementLayout::Layout(nullptr));
+        layout->setHalfSize(HW, HH);
+        layout->setSlope(mirror->alpha());
+        auto pos = _element->owner()->position(_element);
+        if (pos == ElementOwner::PositionAtLeft)
+            layout->paintMode = mirror->theta() > 0
+                ? AxiconElementLayout::PlanoConcaveMirror
+                : AxiconElementLayout::PlanoConvexMirror;
+        else if (pos == ElementOwner::PositionAtRight)
+            layout->paintMode = mirror->theta() > 0
+                ? AxiconElementLayout::ConcavePlanoMirror
+                : AxiconElementLayout::ConvexPlanoMirror;
+        else
+            layout->paintMode = mirror->theta() > 0
+                ? AxiconElementLayout::PlanoConvexLens
+                : AxiconElementLayout::PlanoConcaveLens;
+        layout->init();
+    }
+
+    ELEMENT_LAYOUT_PAINT {
+        if (layout) layout->paint(painter, nullptr, nullptr);
+    }
+}
+
+#endif // SCHEMA_LAYOUT_AXICON_H
